cleaning settings - using disable_print decorator
